refactor(webapi): split client responsibilities
