+++ internal/pixiv/webapi/client.go
@@ -1,49 +1,20 @@
 package webapi
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"io"
 	"net/http"
 	"net/url"
-	"path"
 	"strconv"
 	"strings"
-	"time"
 
 	"github.com/FlanChanXwO/pixiv-cli/internal/pixiv/model"
 	"github.com/FlanChanXwO/pixiv-cli/internal/pixiv/protocol"
-	"github.com/FlanChanXwO/pixiv-cli/internal/utils/text"
 )
 
 const (
 	DefaultWebBase = protocol.WebAPIBase
 	defaultUA      = protocol.WebUserAgent
-	// 两个值是对应 Pixiv Web endpoint 的固定 wire batch 边界；边界测试
-	// 分别以 59/60 与 49/50 锁定换页行为，不是本地结果条数限制。
-	artworkSearchPageSize = 60
-	illustRankingPageSize = 50
 )
-
-// ErrMalformedResponse 标识成功 HTTP 响应无法构成约定 JSON，不包含原始响应体。
-var ErrMalformedResponse = protocol.ErrMalformedResponse
-
-// ErrUnrepresentablePagination 标识 cursor offset 无法安全换算为 Web 页码和下一页边界。
-var ErrUnrepresentablePagination = errors.New("web api pagination cannot represent cursor offset")
-
-// EnvelopeError 保留内部兼容名称；message 不得越过 Web adapter。
-type EnvelopeError = protocol.Failure
-
-// IllustPagesError 保留匿名详情流程中 pages 子阶段，供 facade 精确标注 operation。
-type IllustPagesError struct {
-	err error
-}
-
-func (e *IllustPagesError) Error() string { return fmt.Sprintf("web illust pages: %v", e.err) }
-func (e *IllustPagesError) Unwrap() error { return e.err }
 
 type Client struct {
 	httpClient *http.Client
@@ -263,529 +234,4 @@ func (c *Client) UgoiraMetadata(ctx context.Context, id int64) (*model.UgoiraMet
 		})
 	}
 	return &result, nil
-}
-
-func (c *Client) getJSON(ctx context.Context, path string, query url.Values, out any) error {
-	rawURL, err := c.webURL(path, query)
-	if err != nil {
-		return protocol.Transport(err)
-	}
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, rawURL, nil)
-	if err != nil {
-		return protocol.Transport(err)
-	}
-	c.setHeaders(req)
-	resp, err := c.httpClient.Do(req)
-	if err != nil {
-		return protocol.Transport(err)
-	}
-	defer resp.Body.Close()
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		// reader 失败可能来自底层 transport；不能让其 URL 或代理诊断跨越
-		// Web adapter，统一转换为 protocol 的脱敏失败。
-		return protocol.Transport(err)
-	}
-	if resp.StatusCode < 200 || resp.StatusCode >= 300 {
-		return protocol.HTTPStatus(resp.StatusCode)
-	}
-	if len(bytes.TrimSpace(body)) == 0 {
-		return protocol.MalformedResponse()
-	}
-	if err := json.Unmarshal(body, out); err != nil {
-		return protocol.MalformedResponse()
-	}
-	return nil
-}
-
-func (c *Client) webURL(path string, query url.Values) (string, error) {
-	base, err := url.Parse(c.webBase)
-	if err != nil {
-		return "", err
-	}
-	rel, err := url.Parse(path)
-	if err != nil {
-		return "", err
-	}
-	resolved := base.ResolveReference(rel)
-	if len(query) > 0 {
-		resolved.RawQuery = query.Encode()
-	}
-	return resolved.String(), nil
-}
-
-func (c *Client) setHeaders(req *http.Request) {
-	for key, value := range protocol.WebHeaders() {
-		if key == "User-Agent" && c.userAgent != "" {
-			value = c.userAgent
-		}
-		req.Header.Set(key, value)
-	}
-}
-
-type APIError = protocol.Failure
-
-func webEnvelopeError(string) error {
-	return protocol.UpstreamRejected()
-}
-
-type webPagination struct {
-	page       int
-	nextOffset int
-}
-
-func checkedWebPagination(offset, pageSize int) (webPagination, error) {
-	if offset < 0 || pageSize <= 0 {
-		return webPagination{}, ErrUnrepresentablePagination
-	}
-	quotient := offset / pageSize
-	maxInt := int(^uint(0) >> 1)
-	if quotient == maxInt {
-		return webPagination{}, ErrUnrepresentablePagination
-	}
-	page := quotient + 1
-	if page > maxInt/pageSize {
-		return webPagination{}, ErrUnrepresentablePagination
-	}
-	// Pixiv Web 使用固定批次页码；同时预检下一页边界，避免响应后产生负 cursor。
-	return webPagination{page: page, nextOffset: page * pageSize}, nil
-}
-
-func trimWebPageOffset[T any](items []T, offset, pageSize int) []T {
-	if offset <= 0 || pageSize <= 0 {
-		return items
-	}
-	skip := offset % pageSize
-	if skip >= len(items) {
-		return nil
-	}
-	return items[skip:]
-}
-
-func webHasNext(offset, rawCount int, total int64, pageSize int) bool {
-	if rawCount == 0 {
-		return false
-	}
-	if total > 0 {
-		pageStart := int64(offset - offset%pageSize)
-		if pageStart >= total {
-			return false
-		}
-		// 用减法表达 pageStart+rawCount<total，避免靠近 MaxInt 时加法回绕。
-		return int64(rawCount) < total-pageStart
-	}
-	// 无 total 时，只在完整上游批次后继续；下一空批次会自然收敛。
-	return rawCount >= pageSize
-}
-
-func webSearchOrder(sort string) string {
-	switch sort {
-	case "", string(model.SortModeDateDesc):
-		return "date_d"
-	case string(model.SortModeDateAsc):
-		return "date"
-	default:
-		return sort
-	}
-}
-
-func webSearchMode(target string) string {
-	switch target {
-	case "", string(model.SearchTargetPartialMatchForTags):
-		return "s_tag"
-	case string(model.SearchTargetExactMatchForTags):
-		return "s_tag_full"
-	case string(model.SearchTargetTitleAndCaption):
-		return "s_tc"
-	default:
-		return target
-	}
-}
-
-func webRankingMode(mode string) string {
-	switch mode {
-	case "", string(model.RankingModeDay):
-		return "daily"
-	case string(model.RankingModeDayMale):
-		return "male"
-	case string(model.RankingModeDayFemale):
-		return "female"
-	case string(model.RankingModeWeek):
-		return "weekly"
-	case string(model.RankingModeWeekOriginal):
-		return "original"
-	case string(model.RankingModeWeekRookie):
-		return "rookie"
-	case string(model.RankingModeMonth):
-		return "monthly"
-	default:
-		return mode
-	}
-}
-
-func setDuration(q url.Values, duration string) error {
-	if duration == "" {
-		return nil
-	}
-	now := time.Now()
-	var start time.Time
-	switch duration {
-	case "within_last_day":
-		start = now.AddDate(0, 0, -1)
-	case "within_last_week":
-		start = now.AddDate(0, 0, -7)
-	case "within_last_month":
-		start = now.AddDate(0, -1, 0)
-	default:
-		return fmt.Errorf("pixiv web fallback does not support duration %q", duration)
-	}
-	q.Set("scd", start.Format("2006-01-02"))
-	q.Set("ecd", now.Format("2006-01-02"))
-	return nil
-}
-
-func mapSearchIllust(item webSearchIllust) model.Illust {
-	tags := make([]model.Tag, 0, len(item.Tags))
-	for _, tag := range item.Tags {
-		tags = append(tags, model.Tag{Name: tag})
-	}
-	imageURLs := model.ImageURLs{SquareMedium: item.URL, Medium: item.URL}
-	return model.Illust{
-		ID:        int64(item.ID),
-		Title:     item.Title,
-		Type:      illustType(int(item.IllustType)),
-		PageCount: int(item.PageCount),
-		XRestrict: int(item.XRestrict),
-		User: model.User{
-			ID:      int64(item.UserID),
-			Name:    item.UserName,
-			Account: strconv.FormatInt(int64(item.UserID), 10),
-		},
-		Tags:      tags,
-		ImageURLs: imageURLs,
-		AIType:    int(item.AIType),
-	}
-}
-
-func mapDetailIllust(item webIllustDetail, metaPages []model.MetaPage) model.Illust {
-	tags := mapDetailTags(item.Tags.Tags)
-	imageURLs := mapDetailURLs(item.URLs)
-	if len(metaPages) > 0 {
-		imageURLs = firstImageURLs(imageURLs, metaPages[0].ImageURLs)
-	}
-	pageCount := int(item.PageCount)
-	if pageCount == 0 && len(metaPages) > 0 {
-		pageCount = len(metaPages)
-	}
-	illust := model.Illust{
-		ID:             int64(firstFlexInt64(item.ID, item.IllustID)),
-		Title:          text.FirstNonEmpty(item.Title, item.IllustTitle),
-		Type:           illustType(int(item.IllustType)),
-		PageCount:      pageCount,
-		TotalBookmarks: int(item.BookmarkCount),
-		TotalView:      int(item.ViewCount),
-		XRestrict:      int(item.XRestrict),
-		User: model.User{
-			ID:      int64(item.UserID),
-			Name:    item.UserName,
-			Account: strconv.FormatInt(int64(item.UserID), 10),
-		},
-		Tags:       tags,
-		ImageURLs:  imageURLs,
-		MetaPages:  metaPages,
-		AIType:     int(item.AIType),
-		CreateDate: item.CreateDate,
-		Width:      int(item.Width),
-		Height:     int(item.Height),
-	}
-	if len(metaPages) == 1 {
-		illust.MetaSinglePage.OriginalImageURL = metaPages[0].ImageURLs.Original
-	}
-	return illust
-}
-
-func imageExtension(rawURL string) string {
-	parsed, err := url.Parse(rawURL)
-	if err != nil {
-		return ""
-	}
-	return strings.TrimPrefix(path.Ext(parsed.Path), ".")
-}
-
-func mapRankingIllust(item webRankingItem) model.Illust {
-	tags := make([]model.Tag, 0, len(item.Tags))
-	for _, tag := range item.Tags {
-		tags = append(tags, model.Tag{Name: tag})
-	}
-	return model.Illust{
-		ID:             int64(item.IllustID),
-		Title:          item.Title,
-		Type:           illustType(int(item.IllustType)),
-		PageCount:      int(item.IllustPageCount),
-		TotalBookmarks: int(item.RatingCount),
-		TotalView:      int(item.ViewCount),
-		User: model.User{
-			ID:      int64(item.UserID),
-			Name:    item.UserName,
-			Account: strconv.FormatInt(int64(item.UserID), 10),
-		},
-		Tags: tags,
-		ImageURLs: model.ImageURLs{
-			SquareMedium: item.URL,
-			Medium:       item.URL,
-			Large:        item.URL,
-		},
-	}
-}
-
-func mapDetailTags(items []webTag) []model.Tag {
-	tags := make([]model.Tag, 0, len(items))
-	for _, item := range items {
-		tags = append(tags, model.Tag{Name: item.Tag, TranslatedName: item.Translation.En})
-	}
-	return tags
-}
-
-func mapDetailURLs(urls webDetailURLs) model.ImageURLs {
-	return model.ImageURLs{
-		SquareMedium: text.FirstNonEmpty(urls.Mini, urls.ThumbMini, urls.Small),
-		Medium:       text.FirstNonEmpty(urls.Small, urls.Regular),
-		Large:        text.FirstNonEmpty(urls.Regular, urls.Original),
-		Original:     urls.Original,
-	}
-}
-
-func mapPageURLs(urls webPageURLs) model.ImageURLs {
-	return model.ImageURLs{
-		SquareMedium: urls.ThumbMini,
-		Medium:       text.FirstNonEmpty(urls.Small, urls.Regular),
-		Large:        text.FirstNonEmpty(urls.Regular, urls.Original),
-		Original:     urls.Original,
-	}
-}
-
-func firstImageURLs(primary, fallback model.ImageURLs) model.ImageURLs {
-	return model.ImageURLs{
-		SquareMedium: text.FirstNonEmpty(primary.SquareMedium, fallback.SquareMedium),
-		Medium:       text.FirstNonEmpty(primary.Medium, fallback.Medium),
-		Large:        text.FirstNonEmpty(primary.Large, fallback.Large),
-		Original:     text.FirstNonEmpty(primary.Original, fallback.Original),
-	}
-}
-
-func illustType(value int) string {
-	switch value {
-	case 1:
-		return string(model.IllustTypeManga)
-	case 2:
-		return string(model.IllustTypeUgoira)
-	default:
-		return string(model.IllustTypeIllust)
-	}
-}
-
-func firstFlexInt64(values ...flexInt64) flexInt64 {
-	for _, value := range values {
-		if value != 0 {
-			return value
-		}
-	}
-	return 0
-}
-
-type ajaxEnvelope[T any] struct {
-	Error   bool   `json:"error"`
-	Message string `json:"message"`
-	Body    T      `json:"body"`
-
-	bodyPresent bool
-}
-
-func (e *ajaxEnvelope[T]) UnmarshalJSON(data []byte) error {
-	var wire struct {
-		Error   bool            `json:"error"`
-		Message string          `json:"message"`
-		Body    json.RawMessage `json:"body"`
-	}
-	if err := json.Unmarshal(data, &wire); err != nil {
-		return err
-	}
-	e.Error = wire.Error
-	e.Message = wire.Message
-	body := bytes.TrimSpace(wire.Body)
-	if len(body) == 0 || bytes.Equal(body, []byte("null")) {
-		return nil
-	}
-	if err := json.Unmarshal(body, &e.Body); err != nil {
-		return err
-	}
-	e.bodyPresent = true
-	return nil
-}
-
-type webSearchBody struct {
-	IllustManga webSearchGroup `json:"illustManga"`
-	Illust      webSearchGroup `json:"illust"`
-}
-
-type webSearchGroup struct {
-	Data  requiredWebList[webSearchIllust] `json:"data"`
-	Total flexInt64                        `json:"total"`
-}
-
-type requiredWebList[T any] struct {
-	Items   []T
-	Present bool
-	Valid   bool
-}
-
-func (l *requiredWebList[T]) UnmarshalJSON(data []byte) error {
-	*l = requiredWebList[T]{}
-	l.Present = true
-	if bytes.Equal(bytes.TrimSpace(data), []byte("null")) {
-		return nil
-	}
-	if err := json.Unmarshal(data, &l.Items); err != nil {
-		return err
-	}
-	l.Valid = true
-	return nil
-}
-
-type webSearchIllust struct {
-	ID            flexInt64 `json:"id"`
-	Title         string    `json:"title"`
-	IllustType    flexInt   `json:"illustType"`
-	XRestrict     flexInt   `json:"xRestrict"`
-	AIType        flexInt   `json:"aiType"`
-	URL           string    `json:"url"`
-	Tags          []string  `json:"tags"`
-	UserID        flexInt64 `json:"userId"`
-	UserName      string    `json:"userName"`
-	PageCount     flexInt   `json:"pageCount"`
-	ProfileImgURL string    `json:"profileImageUrl"`
-}
-
-type webIllustDetail struct {
-	ID            flexInt64     `json:"id"`
-	IllustID      flexInt64     `json:"illustId"`
-	Title         string        `json:"title"`
-	IllustTitle   string        `json:"illustTitle"`
-	IllustType    flexInt       `json:"illustType"`
-	XRestrict     flexInt       `json:"xRestrict"`
-	PageCount     flexInt       `json:"pageCount"`
-	UserID        flexInt64     `json:"userId"`
-	UserName      string        `json:"userName"`
-	BookmarkCount flexInt       `json:"bookmarkCount"`
-	ViewCount     flexInt       `json:"viewCount"`
-	AIType        flexInt       `json:"aiType"`
-	CreateDate    string        `json:"createDate"`
-	Width         flexInt       `json:"width"`
-	Height        flexInt       `json:"height"`
-	Tags          webDetailTags `json:"tags"`
-	URLs          webDetailURLs `json:"urls"`
-}
-
-type webDetailTags struct {
-	Tags []webTag `json:"tags"`
-}
-
-type webTag struct {
-	Tag         string `json:"tag"`
-	Translation struct {
-		En string `json:"en"`
-	} `json:"translation"`
-}
-
-type webDetailURLs struct {
-	Mini      string `json:"mini"`
-	ThumbMini string `json:"thumb_mini"`
-	Small     string `json:"small"`
-	Regular   string `json:"regular"`
-	Original  string `json:"original"`
-}
-
-type webPage struct {
-	URLs   webPageURLs `json:"urls"`
-	Width  flexInt     `json:"width"`
-	Height flexInt     `json:"height"`
-}
-
-type webPageURLs struct {
-	ThumbMini string `json:"thumb_mini"`
-	Small     string `json:"small"`
-	Regular   string `json:"regular"`
-	Original  string `json:"original"`
-}
-
-type webRankingResponse struct {
-	Contents  requiredWebList[webRankingItem] `json:"contents"`
-	RankTotal flexInt64                       `json:"rank_total"`
-}
-
-type webRankingItem struct {
-	Title           string    `json:"title"`
-	Tags            []string  `json:"tags"`
-	URL             string    `json:"url"`
-	IllustType      flexInt   `json:"illust_type"`
-	IllustPageCount flexInt   `json:"illust_page_count"`
-	UserName        string    `json:"user_name"`
-	IllustID        flexInt64 `json:"illust_id"`
-	UserID          flexInt64 `json:"user_id"`
-	RatingCount     flexInt   `json:"rating_count"`
-	ViewCount       flexInt   `json:"view_count"`
-}
-
-type webUgoiraMeta struct {
-	OriginalSrc string           `json:"originalSrc"`
-	Src         string           `json:"src"`
-	Frames      []webUgoiraFrame `json:"frames"`
-}
-
-type webUgoiraFrame struct {
-	File  string  `json:"file"`
-	Delay flexInt `json:"delay"`
-}
-
-type flexInt64 int64
-
-func (v *flexInt64) UnmarshalJSON(data []byte) error {
-	data = bytes.TrimSpace(data)
-	if bytes.Equal(data, []byte("null")) || len(data) == 0 {
-		*v = 0
-		return nil
-	}
-	if data[0] == '"' {
-		var text string
-		if err := json.Unmarshal(data, &text); err != nil {
-			return err
-		}
-		if text == "" {
-			*v = 0
-			return nil
-		}
-		parsed, err := strconv.ParseInt(text, 10, 64)
-		if err != nil {
-			return err
-		}
-		*v = flexInt64(parsed)
-		return nil
-	}
-	parsed, err := strconv.ParseInt(string(data), 10, 64)
-	if err != nil {
-		return err
-	}
-	*v = flexInt64(parsed)
-	return nil
-}
-
-type flexInt int
-
-func (v *flexInt) UnmarshalJSON(data []byte) error {
-	var value flexInt64
-	if err := value.UnmarshalJSON(data); err != nil {
-		return err
-	}
-	*v = flexInt(value)
-	return nil
 }
